Unreliable source of randomness Fixed issues 19 (Delete this business block) The get_winning function in treasure contract uses the following parameters as a source of randomness to determine the winning numbers when pre-minting NFTs:
Block time
Block height
Position of the transaction in the block
Although those parameters are volatile and make it harder to guess the winning numbers, they are not a reliable source of randomness from a security perspective.

It's recommended to use an oracle that provides a safe and secure entropy source, e.g.: Nois.